% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customDataGen.R
\name{customDataGen}
\alias{customDataGen}
\title{Custom Data Generator}
\usage{
customDataGen(
  df,
  xvar = NULL,
  yvar = NULL,
  regression_type = "linear",
  success_level = NULL,
  degree = NULL,
  span = 0.75,
  log_y = F,
  log_base = NULL,
  conf_int = F
)
}
\arguments{
\item{df}{An R data frame containing the input data.}

\item{xvar}{The name of the x variable as a string. If null is provided will use first column of dataframe (default: NULL)}

\item{yvar}{The name of the y variable as a string. If null is provided will use second column of dataframe. (default: NULL)}

\item{regression_type}{Type of regression data to generate. Options include "linear", "polynomial", "logistic" (only for binary logistic regression), or "loess". (Default: "linear")}

\item{success_level}{Which level of binary categorical variable should be considered success if using "logistic" regression. If NULL uses alphabetical order. (Default: NULL)}

\item{degree}{The degree for a polynomial or loess regression. If chosen in regression_type. For loess 'degree' must be 0, 1 or 2. (Default: 2 for polynomial, 1 for loess)}

\item{span}{The span for a loess regression. (Default: 0.75)}

\item{log_y}{Specify whether to apply a logarithmic transformation to y for the fitted line for when using a non-linear scale in the \code{drawr()} function. Currently only for linear regression. If TRUE, the fitted line is transformed as log(y) ~ x; if FALSE, the fitted line is not transformed. (default: FALSE)}

\item{log_base}{Log base for log transformation, only applies if log_y is true. If NULL will apply a natural log transformation. Log_base should match log_base choice in \code{drawr()} function. (default: NULL)}

\item{conf_int}{If a 95\% confidence interval should be generated for \code{drawr()} function. Currently only for linear regression. (default: FALSE)}
}
\value{
A list containing the point data and line data (with equation info) processed from inputted data frame.
}
\description{
\code{customDataGen()} takes in an R data frame and processes it to generate data suitable for the
\code{drawr()} function.
}
