% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawr.R
\name{drawr}
\alias{drawr}
\title{Drawr}
\usage{
drawr(
  data,
  linear = "true",
  draw_start = NULL,
  points_end = NULL,
  x_by = 0.25,
  free_draw = T,
  points = "partial",
  aspect_ratio = 1.5,
  title = "",
  x_range = NULL,
  y_range = NULL,
  x_lab = "",
  y_lab = "",
  drawn_line_color = "steelblue",
  data_tab1_color = "steelblue",
  x_axis_buffer = 0.01,
  y_axis_buffer = 0.05,
  show_finished = T,
  shiny_message_loc = NULL
)
}
\arguments{
\item{data}{The data containing line data and point data.}

\item{linear}{Whether the data represents a linear relationship (default: "true").}

\item{draw_start}{The starting point for drawing (default: NULL).}

\item{points_end}{The ending point for the drawn line (default: NULL).}

\item{x_by}{The increment value for x. (default: 0.25)}

\item{free_draw}{Whether to allow freehand drawing (default: T).}

\item{points}{The type of points to be displayed (default: "partial").}

\item{aspect_ratio}{The aspect ratio of the plot (default: 1.5).}

\item{title}{The title of the plot. (default: "")}

\item{x_range}{The range of x values. (default: NULL)}

\item{y_range}{The range of y values.(default: NULL)}

\item{x_lab}{The x-axis label. (default: "")}

\item{y_lab}{The y-axis label. (default: "")}

\item{drawn_line_color}{The color of the drawn line. (default: "steelblue")}

\item{data_tab1_color}{The color of . (default: "steelblue")}

\item{x_axis_buffer}{The buffer for the x-axis. (default: 0.01)}

\item{y_axis_buffer}{The buffer for the y-axis. (default: 0.05)}

\item{show_finished}{Whether to show the finished plot (default: TRUE).}

\item{shiny_message_loc}{The location to display the shiny message. (default = NULL)}
}
\description{
Draws an interactive you-draw-it plot, calling the D3.js file.
}
