% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_youdrawit_payload.R
\name{ggplot_youdrawit_payload}
\alias{ggplot_youdrawit_payload}
\title{Take in a ggplot2 graph and extract data for rerendering in D3}
\usage{
ggplot_youdrawit_payload(p)
}
\arguments{
\item{p}{A ggplot2 graph}
}
\value{
A dataframe containing information needed to render a ggplot into D3 through youdrawit functions
}
\description{
Take in a ggplot2 graph and extract data for rerendering in D3
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
          geom_point(size = 2, colour = "magenta") +
          labs(x = "Weight", y = "MPG")

ggplot_youdrawit_payload(p)
}
