% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearDataGen.R
\name{linearDataGen}
\alias{linearDataGen}
\title{Linear Data Generation}
\usage{
linearDataGen(
  y_xbar,
  slope,
  sigma,
  points_choice = "full",
  points_end_scale = 1,
  N = 30,
  x_min = 0,
  x_max = 20,
  x_by = 0.25
)
}
\arguments{
\item{y_xbar}{The y value when x is average.}

\item{slope}{The slope of the true line data}

\item{sigma}{The standard deviation of the line data.}

\item{points_choice}{Indicates which type of points to generate for point data. (default: "full")}

\item{points_end_scale}{A scaling factor for determining the end point of the drawn line. (default: 1)}

\item{N}{The number of points to generate. (default: 30)}

\item{x_min}{The minimum x value. (default: 0)}

\item{x_max}{The maximum x value. (default: 20)}

\item{x_by}{The increment value for x. (default: 0.25)}
}
\value{
A list containing the generated point data and line data.
}
\description{
\code{linearDataGen()} generates simulated linear line data and point data 
suitable for the \code{drawr()} function.
}
