% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youdrawitR-package.R
\docType{package}
\name{youdrawitR-package}
\alias{youdrawitR}
\alias{youdrawitR-package}
\title{youdrawitR: A Graphical Testing Tool Designed for Use with ggplot2}
\description{
\figure{logo.png}{options: style='float: right; height: 200px'}
The youdrawitR package provides tools for creating interactive “You Draw It”
visualizations using \emph{ggplot2} and \emph{D3}. Users can convert static plots into
interactive graphics, draw trends, and generate data through direct interaction.
The package supports both guided drawing through \code{drawit()} and flexible,
free-form drawing through \code{sketchit()}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Alisa-Krasilnikov/youdrawitR}
  \item \url{https://alisa-krasilnikov.github.io/youdrawitR/}
  \item Report bugs at \url{https://github.com/Alisa-Krasilnikov/youdrawitR/issues}
}

}
\author{
\strong{Maintainer}: Alisa Krasilnikov \email{alisa.krass@gmail.com} (Maintainer status shared between Alisa Krasilnikov and Dr. Emily Robinson)

Authors:
\itemize{
  \item Dillon Murphy [contributor]
  \item Emily Robinson \email{erobin17@calpoly.edu} (Maintainer status shared between Alisa Krasilnikov and Dr. Emily Robinson) [contributor]
  \item Heike Hofmann
  \item Susan VanderPlas
}

}
\keyword{internal}
