% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customDataGen.R
\name{customDataGen}
\alias{customDataGen}
\title{Custom Data Generator}
\usage{
customDataGen(df, xvar = NULL, yvar = NULL)
}
\arguments{
\item{df}{An R data frame containing the input data.}

\item{xvar}{The name of the x variable as a string. If null is provided will use first column of dataframe (default: NULL)}

\item{yvar}{The name of the y variable as a string. If null is provided will use second column of dataframe. (default: NULL)}
}
\value{
A list containing the point data and line data processed from inputted data frame.
}
\description{
\code{customDataGen()} takes in an R data frame and processes it to generate data suitable for the
\code{drawr()} function.
}
