% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customDataGen.R
\name{customDataGen}
\alias{customDataGen}
\title{Custom Data Generator}
\usage{
customDataGen(df, xvar, yvar)
}
\arguments{
\item{df}{An R data frame containing the input data.}

\item{xvar}{The name of the x variable as a string.}

\item{yvar}{The name of the y variable as a string.}
}
\value{
A list containing the point data and line data processed from inputted data frame.
}
\description{
This function takes in an R data frame and processes it to generate data suitable for the `drawr` function.
}
