% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawit.R
\name{drawit}
\alias{drawit}
\title{Render a ggplot scatterplot as D3 via r2d3}
\usage{
drawit(
  p,
  ...,
  width = NULL,
  height = NULL,
  shiny_message_loc = NULL,
  show_on_finish = FALSE
)
}
\arguments{
\item{p}{A ggplot2 graph}

\item{...}{Additional arguments passed to r2d3()}

\item{width}{The width of the desired D3 graph}

\item{height}{The height of the desired D3 graph}

\item{shiny_message_loc}{The location of your shiny message}
}
\value{
An r2d3 htmlwidget rendering the scatterplot
}
\description{
Render a ggplot scatterplot as D3 via r2d3
}
\examples{
library(ggplot2)

## Basic usage
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point(size = 2, colour = "magenta") +
  labs(x = "Weight", y = "MPG")

drawit(p)

## Pipe-friendly usage
(ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point(size = 2, colour = "magenta") +
  labs(x = "Weight", y = "MPG")) |>
  drawit()
}
