% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawr_app.R
\name{drawr_app}
\alias{drawr_app}
\title{Create a Shiny app for drawing}
\usage{
drawr_app(drawr_output)
}
\arguments{
\item{drawr_output}{The output from the \code{drawr()} function to render the drawing interface.}
}
\value{
A Shiny app object.
}
\description{
\code{shinyApp()} creates a Shiny app for drawing an interactive you-draw-it plot for interactive testing of graphics
}
