% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearDataGen.R
\name{linearDataGen}
\alias{linearDataGen}
\title{Linear Data Generation}
\usage{
linearDataGen(
  y_int,
  slope,
  sigma,
  points_choice = "full",
  points_end_scale = 1,
  N = 30,
  x_min = 0,
  x_max = 20,
  conf_int = F
)
}
\arguments{
\item{y_int}{The y intercept of the line data.}

\item{slope}{The slope of the true line data}

\item{sigma}{The standard deviation of the line data.}

\item{points_choice}{Indicates which type of points to generate for point data. (default: "full")}

\item{points_end_scale}{A scaling factor for determining the end point of the drawn line. (default: 1)}

\item{N}{The number of points to generate. (default: 30)}

\item{x_min}{The minimum x value. (default: 0)}

\item{x_max}{The maximum x value. (default: 20)}

\item{conf_int}{If a 95\% confidence interval should be generated for \code{drawr()} function. (default: FALSE)}
}
\value{
A list containing the generated point data and line data (with equation info).
}
\description{
\code{linearDataGen()} generates simulated linear line data (with equation info) and point data 
suitable for the \code{drawr()} function.
}
