% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchit.R
\name{sketchit}
\alias{sketchit}
\title{Render a ggplot scatterplot as D3 via r2d3}
\usage{
sketchit(
  p,
  ...,
  width = NULL,
  height = NULL,
  shiny_message_loc = NULL,
  color_options = TRUE,
  starting_color = NULL,
  palette = NULL,
  stroke_width = 2,
  button_position = c(1, 1)
)
}
\arguments{
\item{p}{A ggplot2 graph}

\item{...}{Additional arguments passed to r2d3()}

\item{width}{The width of the desired D3 graph}

\item{height}{The height of the desired D3 graph}

\item{shiny_message_loc}{The location of your shiny message}

\item{color_options}{Whether or not you want the pallette to show up with multiple options}

\item{starting_color}{The first color you want your user to sketch with}

\item{palette}{What colors you want your user to pick from}

\item{stroke_width}{The widths of the strokes}

\item{button_position}{Where you want the buttons to live on the graph}
}
\value{
An r2d3 htmlwidget rendering the plot
}
\description{
Render a ggplot scatterplot as D3 via r2d3
}
\examples{
library(ggplot2)

## Basic usage
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point(size = 2, colour = "magenta") +
  labs(x = "Weight", y = "MPG")

sketchit(p)

## Pipe-friendly usage
(ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point(size = 2, colour = "magenta") +
  labs(x = "Weight", y = "MPG")) |>
  sketchit()
}
